requirements.txt
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
Deprecated==1.2.13
idna==3.4
packaging==21.3
pycparser==2.21
pydantic==1.10.2
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyzenhub==0.3.1
requests==2.28.1
types-requests==2.28.11.4
types-urllib3==1.26.25.3
typing_extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
